+++ bot.py
@@ -1,4 +1,4 @@
-__version__ = "3.6.3(1)"
+__version__ = "3.6.3(2)"
 
 import os
 import logging